Add files via upload
My first project where I created a predictive model on Python using Jupyter. The data set consists of different types of crops and their requirements, the aim of the project is to create a model from this dataset that can accurately predict which crop to grow in an environment based on the nutrients available in the soil.